trenddy

+++ 0x03-debugging/2-largest_number.c
@@ -12,10 +12,10 @@
  *      * Return: largest number
  *
  */
-double largest_number(double a, double b, double c)
+int largest_number(int a, int b, int c)
 
 {
-	double largest;
+	int largest;
 
 	if ((a >= b) && (a >= c))
 	{
