+++ 0x03-debugging/main.h
@@ -3,6 +3,6 @@
 #include <stdio.h>
 void print_remaining_days(int month, int day, int year);
 int convert_day(int month, int day);
-double largest_number(double, double, double);
+int largest_number(int, int, int);
 void positive_or_negative(int i);
 #endif
